@@ -51,9 +51,8 @@
     "imagemin-svgo": "^6.0.0",
     "markdown-toc": "^1.2.0",
     "mkdirp": "^0.5.1",
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4",
-    "react-dom-factories": "^1.0.1",
+    "react": "^16.3.2",
+    "react-dom": "^16.3.2",
     "remarkable": "^1.7.1",
     "request": "^2.81.0",
     "shelljs": "^0.7.8",
