@@ -63,10 +63,7 @@ class Head extends React.Component {
         )}
         {this.props.config.noIndex && <meta name="robots" content="noindex" />}
         {this.props.redirect && (
-          <meta
-            http-equiv="refresh"
-            content={'0; URL=' + this.props.redirect}
-          />
+          <meta httpEquiv="refresh" content={'0; URL=' + this.props.redirect} />
         )}
         <link
           rel="shortcut icon"
